Move board counters classes to separate module
